added DSA workflow test with page capabilities
added a test that goes through the process of signing up a new fake user while skipping the email verification step, logs in selecting the remember Me checkbox, and then inputs user information to set up data for themself. The README file was updated along with a few minor tweaks to the login page.